雑にファイルテスト. ディレクトリ有無も見て mkdir した方が良さげ

+++ create_sym_link.sh
@@ -1,8 +1,17 @@
 # screen を起動時に立ち上げたい
-ln -s $HOME/dotfiles/.screenrc $HOME/.screenrc
+if [ ! -e $HOME/.screenrc ]; then
+	ln -s $HOME/dotfiles/.screenrc $HOME/.screenrc
+fi
 
 # 以下 .zsh 用。読み込み順が zshenv > zprofile > zshrc > zlogin なので、冒頭の .zshenv で ZDOTDIR を指定してやる
-ln -s $HOME/dotfiles/.zshenv $HOME/.zshenv
-ln -s $HOME/dotfiles/.zsh/.zshrc $HOME/.zsh/.zshrc
-ln -s $HOME/dotfiles/.zsh/.zaliases $HOME/.zsh/.zaliases
+if [ ! -e $HOME/.zshenv ]; then
+	ln -s $HOME/dotfiles/.zshenv $HOME/.zshenv
+fi
 
+if [ ! -e $HOME/.zsh/.zshrc ]; then
+	ln -s $HOME/dotfiles/.zsh/.zshrc $HOME/.zsh/.zshrc
+fi
+
+if [ ! -e $HOME/.zsh/.zaliases ]; then
+	ln -s $HOME/dotfiles/.zsh/.zaliases $HOME/.zsh/.zaliases
+fi
